+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Logins
 
 - Added a disableMemSecurity function to turn off some dubious behaviors of SQLcipher. ([#838](https://github.com/mozilla/application-services/pull/838))
+- The iOS SQLCipher build configuration has been adjusted ([#837](https://github.com/mozilla/application-services/pull/837))
 
 ## Push