+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 **See [the release process docs](docs/howtos/cut-a-new-release.md) for the steps to take when cutting a new release.**
 
-[Full Changelog](https://github.com/mozilla/application-services/compare/v0.21.0...master)
+[Full Changelog](https://github.com/mozilla/application-services/compare/v0.22.0...master)
+
+# v0.22.0 (_2019-03-22_)
+
+[Full Changelog](https://github.com/mozilla/application-services/compare/v0.21.0...v0.22.0)
 
 ## Logins